+++ src/main/java/net/meander/subtlyd/world/item/CreativeModeTabsSD.java
@@ -80,6 +80,7 @@ public class CreativeModeTabsSD {
             entries.insertAfter(itemStack -> itemStack.is(LINGERING_POTION) && itemStack.getOrDefault(DataComponents.POTION_CONTENTS, PotionContents.EMPTY).is(Potions.STRONG_POISON),
                     List.of(PotionContents.createItemStack(LINGERING_POTION, PotionsSD.DECAY)),
                     CreativeModeTab.TabVisibility.PARENT_AND_SEARCH_TABS);
+            entries.insertBefore(POTION, ELIXIR);
         });
 
         CreativeModeTabEvents.modifyOutputEvent(CreativeModeTabs.COMBAT).register(entries -> {